requirements.txt
accelerate==0.23.0
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
bitsandbytes==0.43.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
diffusers==0.14.0
docker-pycreds==0.4.0
einops==0.8.0
eval_type_backport==0.2.0
exceptiongroup==1.2.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
install==1.3.5
ipdb==0.13.13
ipywidgets==8.1.3
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.4
lazy_loader==0.4
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mediapy==1.2.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
open_clip_torch==2.26.1
opencv-python-headless==4.10.0.84
pillow==10.4.0
protobuf==5.27.2
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pytorch-fid==0.3.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
sentry-sdk==2.7.1
setproctitle==1.3.3
smmap==5.0.1
some-package==0.1
sympy==1.12.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==1.0.7
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.4
transformers==4.28.1
triton==2.3.1
urllib3==2.2.2
wandb==0.17.4
widgetsnbextension==4.0.11
zipp==3.19.2
